Macro Int
Add preliminary support for non-vararg macros.
Add integers with some operators, with a test around them.

"int" is an internal type which relies on the "'int" macro for any int
value to be specified. There are three internal functions: add, negate,
and zero. Zero tests to see if the argument is equal to zero.